refactor: add layout to wrap content and move NavigationItem to ui/model
