+++ app/src/main/java/com/acevy/habit_tracker/ui/model/NavigationItem.kt
@@ -1,6 +1,7 @@
-package com.acevy.habit_tracker.ui.navigation
+package com.acevy.habit_tracker.ui.model
 
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.acevy.habit_tracker.ui.navigation.Screen
 
 data class NavigationItem(
     val title: String,
